add:#62 menuのディレクトリ構成修正
